services/external-api/requirements.txt
anyio==3.7.0
azure-common==1.1.28
azure-core==1.29.4
azure-identity==1.14.1
azure-nspkg==3.0.2
azure-storage-blob==12.18.3
certifi==2023.7.22
cffi==2.0.0
charset-normalizer==3.1.0
click==8.3.1
cryptography==41.0.4
exceptiongroup==1.1.1
fastapi==0.97.0
h11==0.14.0
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
portalocker==2.7.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.6
uvicorn==0.22.0
